Data and map syncs and works
We can choose the size and location to simulate on. This is done in
code. Everything else is filled with 0s.
The data sizes all fit and work.
The simulate function has all of the parameters. It just needs to be
connected to the control panel
zooming in on a location and adding an image needs to be added.